backend/requirements.txt
Django==5.2.8
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
psycopg2-binary==2.9.10
qrcode==8.0
Pillow==11.0.0
django-cors-headers==4.6.0
